+++ src/NovaCache/Compression.hs
@@ -9,6 +9,7 @@ module NovaCache.Compression
 where
 
 import qualified Codec.Compression.Lzma as Lzma
+import Control.Exception (SomeException, evaluate, try)
 import Data.ByteString (ByteString)
 import qualified Data.ByteString.Lazy as BL
 
@@ -17,5 +18,11 @@ compressXz :: ByteString -> ByteString
 compressXz = BL.toStrict . Lzma.compress . BL.fromStrict
 
 -- | Decompress an xz-compressed strict 'ByteString'.
-decompressXz :: ByteString -> ByteString
-decompressXz = BL.toStrict . Lzma.decompress . BL.fromStrict
+--
+-- Returns 'Left' with an error message if the input is not valid xz data.
+decompressXz :: ByteString -> IO (Either String ByteString)
+decompressXz bs = do
+  result <- try (evaluate (BL.toStrict (Lzma.decompress (BL.fromStrict bs))))
+  pure $ case result of
+    Left err -> Left ("xz decompression failed: " ++ show (err :: SomeException))
+    Right decompressed -> Right decompressed